thinking of changing to this structure.
